+++ app/Http/Controllers/CampaignController.php
@@ -18,8 +18,8 @@ class CampaignController extends Controller
     public function index(Client $client)
     {
         //
-        //$campaigns = Campaign::where('client_id', '=', $client->id)->get();
-        $campaigns = $client->campaigns()->get();
+        //
+        $campaigns = $client->campaigns->get();
         return CampaignResource::collection($campaigns);
 
     }
@@ -33,13 +33,20 @@ class CampaignController extends Controller
     public function store(Request $request)
     {
         //
-        $campaign = new Campaign;
-        $campaign->name = $request->input('name');
-        $campaign->start_date = date('Y-m-d',$request->input('start_date'));
-        //$campaign->start_date = $request->input('start_date');
-        $campaign->end_date = date('Y-m-d',$request->input('end_date'));
-        $campaign->active = $request->input('active') ? 1 : 0;
-        return new CampaignResource($campaign);
+        $client = \App\Client::find($request->client_id);
+        if($client){
+            $campaign = new \App\Campaign;
+            $campaign->name = $request->name;
+            $campaign->start_date = $request->start_date;
+            $campaign->end_date = $request->end_date;
+            $campaign->active = $request->active;
+            $client->campaigns()->save($campaign);
+            $campaign->save();
+            return new CampaignResource($campaign);
+        } else {
+            return response()->json('No existe el cliente', 400);
+        }
+        
 
     }
 
@@ -51,14 +58,8 @@ class CampaignController extends Controller
      */
     public function show(Client $client, Campaign $campaign)
     {
-        /*
-        $campaign = Campaign::findOrFail($id);
-        if($campaign->client_id == $client->id){
-        return new CampaignResource($campaign);
-        }
-        return response()->json(['error'=>'Resource not found'], 404);
-         */
-        CampaignResource::withoutWrapping();
+        //
+        
         return new CampaignResource($campaign);
     }
 
@@ -69,22 +70,28 @@ class CampaignController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Client $client, Request $request, $id)
+    public function update(Request $request, $id)
     {
         //
-        $campaign = Campaign::findOrFail($id);
-        $campaign->name = $request->input('name');
-        $campaign->start_date = $request->input('start_date');
-        $campaign->end_date = $request->input('end_date');
-        $campaign->active = $request->input('active');
-
-        foreach ($request->input('beacons') as $beacon_id) {
-            $campaign->beacons()->save(Beacon::find($beacon_id));
+        $client = \App\Client::find($campaign->client_id);
+        if($client){
+            $campaign = $client->campaigns()->find($id);
+            if($campaign){
+                $campaign->name = $request->name;
+                $campaign->start_date = $request->start_date;
+                $campaign->end_date = $request->end_date;
+                $campaign->active = $request->active;
+                $client->campaigns()->save($campaign);
+                $campaign->save();
+                return new CampaignResource($campaign);
+            } else {
+                return response()->json('No existe la campaña', 400);
+            }
+        } else {
+            return response()->json('No existe el cliente', 400);
         }
 
-        if ($client->campaigns()->save($campaign)) {
-            return new CampaignResource($campaign->fresh());
-        }
+        
     }
 
     /**
@@ -96,8 +103,11 @@ class CampaignController extends Controller
     public function destroy($id)
     {
         //
-        $campaign = Campaign::findOrFail($id);
-        $campaign->delete();
-        return response()->json('No existe el elemento', 204);
+        $campaign = Campaign::find($id);
+        if($campaign){
+            $campaign->delete();
+            return response()->json('Elemento eliminado', 200);
+        }
+        return response()->json('No existe la campaña', 40);
     }
 }
